Se contemplan que no se pueden borrar coros ni partituras con entidades asociadas

+++ chorus/app/Http/Controllers/CoroController.php
@@ -110,6 +110,11 @@ class CoroController extends Controller
     {
         $coro = Coro::find($id);
         if (isset($coro)) {
+            $partituras = $coro->partituras->count();
+
+            if ($partituras > 0) {
+                return ['Hay ' . $partituras . ' partituras asociadas',''];
+            }
             $res = Coro::destroy($id);
             if ($res) {
                 return 1;
